Add autoPingInterval and autoPingTimeout
This makes the connection more robust. Previously, a dropped connection would not be detected until the client attempts to send a message. For a "read-only" client that only listens to server messages, the dropped connection would never be detected.

With autoPingInterval and autoPingTimeout added, this is no longer an issue. I tested this change in a broadcaster-listener setup using iptables to artificially block packets.

See: http://nicolovaligi.com/automatic-reconnection-websockets-autobahn.html